requirements.txt
meshtastic
paho-mqtt
protobuf
cryptography